[all] remove "comparison between signed and unsigned integer expressions"
* choose of better types for local variables
* use of if (X < 0) {} before using cast to unsigned types
* change of some functions signatures

+++ libpvkernel/src/rush/PVXmlTreeNodeDom.cpp
@@ -1001,7 +1001,7 @@ void PVRush::PVXmlTreeNodeDom::getGroupsByType(types_groups_t& grps)
 		}
 	}
 
-	for (size_t ichild = 0; ichild < getChildren().size(); ichild++) {
+	for (int ichild = 0; ichild < getChildren().size(); ichild++) {
 		getChild(ichild)->getGroupsByType(grps);
 	}
 }
